refactor: searchModal에 event bind로 변경
